requirements.txt
flask>=3.0.0
anthropic>=0.21.0
python-dotenv>=1.0.0
tavily-python>=0.5.0
flask-cors>=4.0.0
pandas>=2.0.0
openpyxl>=3.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
